+++ R/themes.R
@@ -219,8 +219,15 @@ register_legendry_elements <- function() {
     legendry.axis.mini.ticks.length = rel(0.5),
     legendry.guide.spacing = unit(2.25, "pt"),
     legendry.group.spacing = rel(2),
-    legendry.axis.subtitle = element_text(margin = margin(5.5, 5.5, 5.5, 5.5)),
+    legendry.axis.subtitle = NULL,
     legendry.axis.subtitle.position = c("left", "top", "bottom", "right"),
+    legendry.symbol = element_point(),
+    legendry.connector = element_line(),
+    legendry.zebra.light = NULL,
+    legendry.zebra.dark = NULL,
+    legendry.table.vlines = element_blank(),
+    legendry.table.hlines = element_blank(),
+    legendry.table.spacing = rel(0.5),
     element_tree = list(
       legendry.bracket.size = el_def("unit"),
       legendry.bracket = el_def(line, "line"),
@@ -240,7 +247,14 @@ register_legendry_elements <- function() {
       legendry.guide.spacing = el_unit("axis.ticks.length"),
       legendry.group.spacing = el_unit("legend.key.spacing"),
       legendry.axis.subtitle = el_def(text, "axis.text"),
-      legendry.axis.subtitle.position = el_def("character")
+      legendry.axis.subtitle.position = el_def("character"),
+      legendry.symbol = el_def(element_point, "point"),
+      legendry.connector = el_def(line, "line"),
+      legendry.zebra.light = el_def(rect, "plot.background"),
+      legendry.zebra.dark  = el_def(rect, "panel.background"),
+      legendry.table.vlines = el_def(line, "axis.line"),
+      legendry.table.hlines = el_def(line, "axis.line"),
+      legendry.table.spacing = el_def(c("unit", "rel"), "spacing")
     )
   )
 }