+++ NEWS.md
@@ -2,11 +2,17 @@
 
 * New guide `guide_legend_manual()` for user defined keys (#88)
 * New guide `guide_axis_plot()` to place side-plots (#43)
+* New guide `guide_axis_upset()` to place symbols in upset matrices next to axes. 
+  Also comes with `key_upset()` to match (#38).
+* New guide `guide_axis_symbols()` for a more manual version of placing symbols. 
+  Also comes with `key_symbols()` to match (#38).
 * New arguments `guide_axis_base(subtitle)` and `guide_axis_nested(subtitle)` 
   make it easier to title custom axes (#102)
 * New arguments `guide_legend_cross(row_title, col_title, subtitle_position)`
   to display subtitles over dimensions (#79).
 * `compose_stack(side.titles)` also work in non-position guides (#101)
+* Changed the margins and inferred justification of the `legendry.axis.subtitle`
+  and `legendry.legend.subtitle` theme elements.
 * Updated title mechanism to comply with ggplot2#6200 (#104)
 * `key_group_split()` tries to preserve order of labels better, rather than 
   re-sorting them (#100)